python_ocr_service/requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
pytesseract==0.3.10
Pillow==9.4.0
face_recognition==1.3.0

face-recognition~=1.3.0
numpy~=2.2.3